moved db migration in backend_core to db module
